requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.0
google-cloud-bigquery==3.11.0
torch-fidelity==0.3.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2